Decouple monitor and animations.
Delegate drawing to threads as was done earlier. This makes it possible
to access the monitor while animations are playing for adding callers
with the callLift()~ function. While at the same time preventing actions
that would generate more animations from executing.

The earlier mistake was that the "animation lock" was not set inside of
the synchronized context. If you try to set and unset the lock outside
of the critical section, there are going to be race conditions..

Correct way of doing it is to set it in the critical region and the use
the synchronized method in the thread to unset it.